requirements.txt
django-rest-framework-proxy>=1.6.0,<1.7
